project/backend/requirements.txt
fastapi
uvicorn
pydantic
confluent_kafka
google-auth
sqlalchemy
pymysql
pyjwt
requests
cryptography
python-jose